Add initial data structures for accounts and games (#1)
* Add initial data structures for accounts and games

* Proposed Schema Updates

Proposed Schema Updates
===

**TL;DR:**

Proposes several new models and other schema changes so that the
application can carryout all facets of a trivia game.

Proposed Models
---

Proposes adding the following models necessary to play a game of trivia:

* `Category`
* `Game::Player`
* `Game::Player::Answer`
* `Game::Question`
* `Question`
* `Question::Answer`
* `Question::Category`

`Category`
----

Will serve as a catalog of categories sourced from approved trivia providers.

**Tasks**

- [x] Migration
- [ ] Model

`Game::Player`
---

Will keep a record of each player playing in a game.

**Tasks**

- [x] Migration
- [ ] Model

`Game::Player::Answer`
---

Will keep a record of players selected answer to a question in a game

**Tasks**

- [ ] Migration
- [ ] Model

`Game::Question`
---

Will serve as a direct relationship between a game and its questions.

**Tasks**

- [ ] Migration
- [ ] Model

`Question`
---

It Will serve as a bank of questions sourced from approved trivia
providers.  It should also enforce a `uid` or unique identifier per
trivia provider to reduce the level of duplicate questions

 * Will default to a source provided unique identifier.
 * Will fallback to a base64 hash of the question prompt prefixed with
   the trivia provider id.

 ##### EXAMPLE
 ```
 #{trivial_provider_id}-#{base_64_encoded_question_prompt}
 ```

**Tasks**

- [x] Migration
- [ ] Model

`Question::Category`
---

Will serve as a direct relationship between a question and category.

**Tasks**

- [x] Migration
- [ ] Model

`Question::Answer`
---

Will serve as a direct relationship between a question and at least one correct answer.

**Tasks**

- [X] Migration
- [ ] Model

Additional Proposals
---

Discuss the following proposals and document decisions in ADRs

* Rename `Game::Player::Answer` to `Game::Answer`
* Choose hard coded trivia providers?
* Choose `acts_as_list` for sequencing answers to questions

**Rename `Game::Player::Answer` to `Game::Answer`?**

Given the structure of the rest of the game tables, is the direct
relationship to a player implicit enough to warrant the removal of
`Player` in the model namespace?

**Choose hard coded trivia providers?**

Should we start with hard coded provider objects of agreed upon trivia
API providers? What benefits would a provider stored in the database
bring this early in the project?

**Choose `acts_as_list` for sequencing answers to questions**

Do we need to store a specific sequence of answers for questions? If so,
should we use the `acts_as_list` gem to implement this feature?

* Add question, answers, games, and players to the data model

Co-authored-by: Mike Elkins <mike@teamairship.com>